+++ test/ops/test_normalize.py
@@ -2,16 +2,17 @@
 Test for the harmonization operation
 """
 
+import calendar
+from datetime import datetime
 from unittest import TestCase
 
+import numpy as np
 import xarray as xr
 from jdcal import gcal2jd
-import numpy as np
-from datetime import datetime
-import calendar
+from numpy.testing import assert_array_almost_equal
 
-from cate.ops.normalize import normalize, adjust_spatial_attrs, adjust_temporal_attrs
 from cate.core.op import OP_REGISTRY
+from cate.ops.normalize import normalize, adjust_spatial_attrs, adjust_temporal_attrs, fix_lon_360
 from cate.util.misc import object_to_qualified_name
 
 
@@ -106,6 +107,16 @@ class TestNormalize(TestCase):
         actual = normalize(dataset)
         assertDatasetEqual(actual, expected)
 
+    def test_normalize_with_missing_time_dim(self):
+        ds = xr.Dataset({'first': (['lat', 'lon'], np.zeros([90, 180])),
+                         'second': (['lat', 'lon'], np.zeros([90, 180]))},
+                        coords={'lat': np.linspace(-89.5, 89.5, 90),
+                                'lon': np.linspace(-179.5, 179.5, 180)},
+                        attrs={'time_coverage_start': '20120101',
+                               'time_coverage_end': '20121231'})
+        norm_ds = normalize(ds)
+        self.assertIs(norm_ds, ds)
+
     def test_normalize_julian_day(self):
         """
         Test Julian Day -> Datetime conversion
@@ -162,6 +173,7 @@ class TestAdjustSpatial(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['geospatial_lat_min']
 
         # Make sure expected values are in the new dataset
@@ -212,6 +224,7 @@ class TestAdjustSpatial(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['geospatial_lat_min']
 
         # Make sure expected values are in the new dataset
@@ -276,6 +289,7 @@ class TestAdjustSpatial(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['geospatial_lat_min']
 
         # Make sure expected values are in the new dataset
@@ -341,6 +355,7 @@ class TestAdjustSpatial(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['geospatial_lat_min']
 
         # Make sure expected values are in the new dataset
@@ -410,10 +425,9 @@ class TestAdjustSpatial(TestCase):
         ds.lon.attrs['units'] = 'degrees_east'
         ds.lat.attrs['units'] = 'degrees_north'
 
-        with self.assertRaises(ValueError) as cm:
-            adjust_spatial_attrs(ds)
-
-        self.assertEqual(str(cm.exception), 'Cannot determine spatial extent for dimension "lon"')
+        ds2 = adjust_spatial_attrs(ds)
+        # Datasets should be the same --> not modified
+        self.assertIs(ds2, ds)
 
 
 class TestAdjustTemporal(TestCase):
@@ -429,6 +443,7 @@ class TestAdjustTemporal(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_start']
 
         # Make sure expected values are in the new dataset
@@ -442,6 +457,7 @@ class TestAdjustTemporal(TestCase):
                          'P336D')
 
         # Test existing attributes update
+        # noinspection PyTypeChecker
         indexers = {'time': slice(datetime(2000, 2, 15), datetime(2000, 6, 15))}
         ds2 = ds1.sel(**indexers)
         ds2 = adjust_temporal_attrs(ds2)
@@ -480,6 +496,7 @@ class TestAdjustTemporal(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_start']
 
         # Make sure expected values are in the new dataset
@@ -510,6 +527,7 @@ class TestAdjustTemporal(TestCase):
 
         # Make sure original dataset is not altered
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_start']
 
         # Make sure expected values are in the new dataset
@@ -521,6 +539,7 @@ class TestAdjustTemporal(TestCase):
                          'P31D')
         with self.assertRaises(KeyError):
             # Resolution is not defined for a single slice
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_resolution']
 
         # Without bnds
@@ -538,6 +557,76 @@ class TestAdjustTemporal(TestCase):
         self.assertEqual(ds1.attrs['time_coverage_end'],
                          '2000-01-01T00:00:00.000000000')
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_resolution']
         with self.assertRaises(KeyError):
+            # noinspection PyStatementEffect
             ds.attrs['time_coverage_duration']
+
+    def test_time_and_time_bnds_dim_generated(self):
+        ds = xr.Dataset({'first': (['lat', 'lon'], np.zeros([90, 180])),
+                         'second': (['lat', 'lon'], np.zeros([90, 180]))},
+                        coords={'lat': np.linspace(-89.5, 89.5, 90),
+                                'lon': np.linspace(-179.5, 179.5, 180)},
+                        attrs={'time_coverage_start': '20120101',
+                               'time_coverage_end': '20121231'})
+
+        new_ds = adjust_temporal_attrs(ds)
+
+        self.assertIsNot(ds, new_ds)
+        self.assertEqual(len(new_ds.coords), 4)
+        self.assertIn('lon', new_ds.coords)
+        self.assertIn('lat', new_ds.coords)
+        self.assertIn('time', new_ds.coords)
+        self.assertIn('time_bnds', new_ds.coords)
+
+        import pandas as pd
+        self.assertEqual(new_ds.first.shape, (1, 90, 180))
+        self.assertEqual(new_ds.second.shape, (1, 90, 180))
+        self.assertEqual(new_ds.coords['time'][0], xr.DataArray(pd.to_datetime('2012-07-01T12:00:00')))
+        self.assertEqual(new_ds.coords['time_bnds'][0][0], xr.DataArray(pd.to_datetime('2012-01-01')))
+        self.assertEqual(new_ds.coords['time_bnds'][0][1], xr.DataArray(pd.to_datetime('2012-12-31')))
+
+    def test_only_time_dim_generated(self):
+        ds = xr.Dataset({'first': (['lat', 'lon'], np.zeros([90, 180])),
+                         'second': (['lat', 'lon'], np.zeros([90, 180]))},
+                        coords={'lat': np.linspace(-89.5, 89.5, 90),
+                                'lon': np.linspace(-179.5, 179.5, 180)},
+                        attrs={'time_coverage_start': '20120101'})
+
+        new_ds = adjust_temporal_attrs(ds)
+
+        self.assertIsNot(ds, new_ds)
+        self.assertEqual(len(new_ds.coords), 3)
+        self.assertIn('lon', new_ds.coords)
+        self.assertIn('lat', new_ds.coords)
+        self.assertIn('time', new_ds.coords)
+        self.assertNotIn('time_bnds', new_ds.coords)
+
+        import pandas as pd
+        self.assertEqual(new_ds.first.shape, (1, 90, 180))
+        self.assertEqual(new_ds.second.shape, (1, 90, 180))
+        self.assertEqual(new_ds.coords['time'][0], xr.DataArray(pd.to_datetime('2012-01-01')))
+
+
+class TestFix360Lon(TestCase):
+    def test_fix_360_lon(self):
+        lon_size = 360
+        lat_size = 130
+        time_size = 12
+        ds = xr.Dataset({
+            'first': (['time', 'lat', 'lon'], np.random.random_sample([time_size, lat_size, lon_size])),
+            'second': (['time', 'lat', 'lon'], np.random.random_sample([time_size, lat_size, lon_size])),
+            'lon': np.linspace(1., 360., lon_size),
+            'lat': np.linspace(-65, 65, lat_size),
+            'time': [datetime(2000, x, 1) for x in range(1, time_size + 1)]})
+
+        new_ds = fix_lon_360(ds)
+        self.assertIsNot(ds, new_ds)
+        self.assertEqual(ds.dims, new_ds.dims)
+        self.assertEqual(ds.sizes, new_ds.sizes)
+        assert_array_almost_equal(new_ds.lon, np.linspace(-179.5, 179.5, 360))
+        assert_array_almost_equal(new_ds.first[..., :180], ds.first[..., 180:])
+        assert_array_almost_equal(new_ds.first[..., 180:], ds.first[..., :180])
+        assert_array_almost_equal(new_ds.second[..., :180], ds.second[..., 180:])
+        assert_array_almost_equal(new_ds.second[..., 180:], ds.second[..., :180])