+++ CHANGES.md
@@ -1,7 +1,13 @@
 ## Version 2.0.0.dev10 (in development)
 
+* Support datasets with 0,360 degree longitude ranges [#620](https://github.com/CCI-Tools/cate/issues/620)
+* Temporal aggregation operation can now aggregate to pre-defined seasons, as well as custom resolutions [#472](https://github.com/CCI-Tools/cate/issues/472)
+* We now use "MB" units instead of "MiB" (part of [#325](https://github.com/CCI-Tools/cate/issues/325))
 * Fixed a bug with animation generation [#585](https://github.com/CCI-Tools/cate/issues/585)
 * Upgrade to using newer xarray version after an upstream bugfix [#579](https://github.com/CCI-Tools/cate/issues/579)
+* Fixed a bug of unable to do pixel values extraction if one of the workflow has an error [#616](https://github.com/CCI-Tools/cate/issues/616)
+* Add the capability to create Hovmoeller plots [#503](https://github.com/CCI-Tools/cate/issues/503)
+* Add a reduce operation that lets the user create arbitrary data reductions [#618](https://github.com/CCI-Tools/cate/issues/618)
 
 ## Version 2.0.0.dev9